+++ browser-extension/mv3/test-results/.last-run.json
@@ -1,4 +1,4 @@
 {
-  "status": "failed",
-  "failedTests": ["fac79d7cf71249642fdf-98cadbd4ef48a564a740"]
+  "status": "passed",
+  "failedTests": []
 }
