+++ browser-extension/mv3/tests/rules/redirect/redirectRule.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from "@playwright/test";
-import { test } from "../fixtures";
-import { loadRules } from "../utils";
+import { test } from "../../fixtures";
+import { loadRules } from "../../utils";
 import redirectRules from "./redirect_rules.json";
 
 test.describe("Redirect Rule", () => {
