+++ browser-extension/mv3/tests/rules/redirect/redirect_rules.json
